Rename example file
